move json data to seperate files
