+++ packages/db/tsconfig.json
@@ -40,7 +40,7 @@
         // "resolvePackageJsonImports": true,                /* Use the package.json 'imports' field when resolving imports. */
         // "customConditions": [],                           /* Conditions to set in addition to the resolver-specific defaults when resolving imports. */
         // "noUncheckedSideEffectImports": true,             /* Check side effect imports. */
-        // "resolveJsonModule": true,                        /* Enable importing .json files. */
+        "resolveJsonModule": true /* Enable importing .json files. */,
         // "allowArbitraryExtensions": true,                 /* Enable importing files with any extension, provided a declaration file is present. */
         // "noResolve": true,                                /* Disallow 'import's, 'require's or '<reference>'s from expanding the number of files TypeScript should add to a project. */
